do some refine
